ref(dashboards): Extract shared applyDashboardFilters utility
Move duplicate applyDashboardFilters implementations from 9 widget query
hooks into a single shared function in genericWidgetQueries.tsx. This
reduces code duplication and ensures consistent behavior across all
dataset-specific widget hooks.
